Fix silent test failures for Chromium Gtests
Previously, we would depend on the test parser to call handle the
lifecycle of the test listener (eg: call testRunStarted, testPassed,
testFailed, ...). However, this was buggy because any failure that
happened before the parser wouldn't call `testRunStarted` which leads TH
to believe that the test passed since nothing was started.

The current approach is to make sure that testRunStarted is always
called before the return the control to the parent runner.

Deleted FailedChromiumGTestException as it served no purpose, and replaced its occurrences with IllegalStateException.

Bug: 402068773
Change-Id: Iebebd277a1ba43d227b935858c8dcc20c0044247