Import Cronet v136.0.7091.2 from Chromium to AOSP.
This CL was generated by the following steps:

1. git fetch --tags && git checkout 136.0.7091.2
2. git fetch https://chromium.googlesource.com/chromium/src refs/changes/75/6393675/4 && git cherry-pick FETCH_HEAD
3. git fetch https://chromium.googlesource.com/chromium/src refs/changes/35/6393635/6 && git cherry-pick FETCH_HEAD
4. vpython3 components/cronet/gn2bp/run_gn2bp.py --git-url-and-branch file:////path/to/android-main/external/cronet branch_name --channel stable

Merge conflicts have been resolved in patchset (2).
The consistency file is regenerated in patchset (3).
Additional Android.bp changes in patchset (4).

Patchset(4):

https://crrev.com/c/6219368 added a new target as the main final target
that should be linked into the binaries. However, our tests maually
select the JNI registration targets which resulted in a failure to find
the appropriate Soong module.

Bug: 414389515
Change-Id: I9566c04092988337a162eb95a09f9357489bbfd6